fix: move useTree hook from frontend to resolver
