+++ libs/frontend/modules/admin/src/use-cases/execute-command/ExecuteCommandButton.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import { useAdminDispatch } from '../../hooks'
 
 export const ExecuteCommandButton = () => {
-  const { openExecuteCommandModal } = useAdminDispatch()
+  // const { openExecuteCommandModal } = useAdminDispatch()
 
-  return <Button onClick={openExecuteCommandModal}>Execute Command</Button>
+  return <Button>Execute Command</Button>
 }
