+++ libs/frontend/modules/admin/src/use-cases/execute-command/ExecuteCommandModal.tsx
@@ -11,34 +11,35 @@ import { useExecuteCommandForm } from './useExecuteCommandForm'
  * @constructor
  */
 export const ExecuteCommandModal = () => {
-  const {
-    onSubmit,
-    actionType,
-    onSubmitSuccess,
-    onSubmitError,
-    reset,
-    isLoading,
-  } = useExecuteCommandForm()
+  // const {
+  //   onSubmit,
+  //   actionType,
+  //   onSubmitSuccess,
+  //   onSubmitError,
+  //   reset,
+  //   isLoading,
+  // } = useExecuteCommandForm()
 
   return (
-    <FormModal
-      okButtonProps={{ loading: isLoading }}
-      okText="Execute Command"
-      onCancel={reset}
-      visible={actionType === ExecuteCommandActionType.Execute}
-    >
-      {({ submitRef }) => (
-        <Form<ExecuteCommandInput, ExecuteCommandResponse>
-          model={{}}
-          onSubmit={onSubmit}
-          onSubmitError={onSubmitError}
-          onSubmitSuccess={onSubmitSuccess}
-          schema={executeCommandSchema}
-          submitRef={submitRef}
-        >
-          <AutoFields />
-        </Form>
-      )}
-    </FormModal>
+    <div></div>
+    // <FormModal
+    //   okButtonProps={{ loading: isLoading }}
+    //   okText="Execute Command"
+    //   onCancel={reset}
+    //   visible={actionType === ExecuteCommandActionType.Execute}
+    // >
+    //   {({ submitRef }) => (
+    //     <Form<ExecuteCommandInput, ExecuteCommandResponse>
+    //       model={{}}
+    //       onSubmit={onSubmit}
+    //       onSubmitError={onSubmitError}
+    //       onSubmitSuccess={onSubmitSuccess}
+    //       schema={executeCommandSchema}
+    //       submitRef={submitRef}
+    //     >
+    //       <AutoFields />
+    //     </Form>
+    //   )}
+    // </FormModal>
   )
 }
