+++ libs/frontend/modules/admin/src/use-cases/export/ExportButton.tsx
@@ -1,5 +1,5 @@
 import { useGetAtomsQuery } from '@codelab/frontend/modules/atom'
-import { useGetTagGraphsQuery, useTagTree } from '@codelab/frontend/modules/tag'
+import { useGetTagGraphsQuery } from '@codelab/frontend/modules/tag'
 import {useGetAllTypesQuery} from '@codelab/frontend/modules/type'
 import { Button } from 'antd'
 import fileDownload from 'js-file-download'
@@ -10,22 +10,8 @@ import sanitizeFilename from 'sanitize-filename'
 export const ExportButton = () => {
   // Tags data
   const { data: tagsData } = useGetTagGraphsQuery()
-  const tagTree = useTagTree(tagsData?.tagGraphs)
-  const rootTagId = tagTree.getRootVertex()?.id
-  console.log('root tag id', rootTagId)
-  let tags: any[] = []
-  if(rootTagId) {
-    tagTree.bfsVisit((v)=> {
-      const parent = v.parent(v.data().id)[0]
-      tags.push({
-        id: v.data().id,
-        name: v.data().name,
-        isRoot: v.data().isRoot,
-        parent: parent ? {id: parent.data().id, name: parent.data().name, isRoot: parent.data().isRoot} : undefined
-      })
-    }, rootTagId)
-  }
-  
+  const tags = tagsData?.tagGraphs ?? { vertices: [], edges: [] }
+
   // All Types Data
   const {data: allData}= useGetAllTypesQuery()
   const allTypes = allData?.types ?? []
